allow unexpected child shards to be ignored
now instead of always throwing an assertion if a child shard has an
open parent, consider worker configuration before doing so.  also, do
not create leases for such shards during shard sync.  this is intended
to mitigate failing worker init when processing dynamodb streams with
many thousands of shards (which can happen for tables with thousands
of partitions).

excludes tests for this new behavior.